refactor: Organize protractor setup
