+++ test/e2e/protractor.conf.js
@@ -1,10 +1,13 @@
-'use strict'
+'use strict';
+
+let chai = require('chai');
+let chaiAsPromised = require('chai-as-promised');
 
 require('babel-register');
 require('babel-polyfill');
 
 exports.config = {
-  specs: 'test/e2e/**/*.js',
+  specs: ['./helper.js', './**/*.test.js'],
   baseUrl: 'http://localhost:3003/',
   directConnect: true,
   seleniumAddress: 'http://localhost:4444/wd/hub',
@@ -13,12 +16,10 @@ exports.config = {
     reporter: 'spec',
     slow: 6000,
     timeout: 10000,
-    compilers: 'js:babel-register'
+    compilers: 'js:babel-register',
   },
   onPrepare: () => {
     browser.ignoreSynchronization = true;
-    let chai = require('chai');
-    let chaiAsPromised = require('chai-as-promised');
 
     chai.use(chaiAsPromised);
     global.expect = chai.expect;
