+++ test/e2e/static-pages/front-page.test.js
@@ -1,24 +1,7 @@
-import fs from 'fs';
-
 describe('Static Front Page', () => {
   beforeEach(() => {
     browser.get('/');
     browser.sleep(1000);
-  });
-
-  // based on https://github.com/angular/protractor/issues/114#issuecomment-29046939
-  afterEach(async function () {
-    let lastTest = this.currentTest;
-
-    if (lastTest.state === 'failed') {
-      let filename = `exception_${lastTest.title}.png`;
-      let png = await browser.takeScreenshot();
-      let buffer = new Buffer(png, 'base64');
-      let stream = fs.createWriteStream(filename);
-
-      stream.write(buffer);
-      stream.end();
-    }
   });
 
   it('shows the front page', async () => {
